增加评论回复列表显示

+++ src/main/java/dzkjdx/jsb/web_community/model/Comment.java
@@ -7,7 +7,7 @@ public class Comment {
     private Long id;
     private Long parentId;
     private Integer type;
-    private Integer commentator;
+    private Long commentator;
     private Long gmtCreate;
     private Long gmtModified;
     private Long likeCount;
